Getting settled in with git, fixing layout.
Got gitignore in, basically a clone of hgignore (which I should probably kill now).
Killed the repository directories in .hgprompt and .hgtemplates. Due to subrepo behavior in Mercurial that I wasn't aware of, these were not showing up in `hg status` and so I had missed them until I moved to git.
Killed executable bits on a few files.